+++ src/setup.boot.ml
@@ -1,0 +1,1 @@
+let library_path = Some []