+++ src/setup.mli
@@ -1,0 +1,9 @@
+(** Setup of Jbuilder *)
+
+(** These parameters are set by [ocaml configure.ml] or copied from
+    [setup.defaults.ml]. During bootstrap, values from [setup.boot.ml]
+    are used *)
+
+(** Where to find installed libraries for the default context. If
+    [None], auto-detect it using standard tools such as [ocamlfind]. *)
+val library_path : string list option